app/requirements.txt
fastapi==0.112.2
uvicorn[standard]==0.30.6
numpy==1.26.4
pillow==10.4.0
tflite-runtime==2.14.0

python-multipart==0.0.9
